requirements.txt
python-telegram-bot==20.7
openai>=1.30.0
python-dotenv==1.0.0
Pillow==10.2.0
aiofiles==23.2.1
